+++ lang/include/ByteArray.hpp
@@ -16,45 +16,45 @@ DECLARE_CLASS(ByteArray) {
 public:
     explicit _ByteArray();
     
-    explicit _ByteArray(uint64_t length);
+    explicit _ByteArray(size_t length);
     
-    explicit _ByteArray(byte *data, uint64_t len,bool mapped = false);
+    explicit _ByteArray(byte *data, size_t len,bool mapped = false);
     
-    explicit _ByteArray(sp<_ByteArray> &, uint64_t start = 0, uint64_t len = 0);
+    explicit _ByteArray(sp<_ByteArray> &, size_t start = 0, size_t len = 0);
 
     ~_ByteArray() override;
 
     byte *toValue(bool copy = false);
     
-    uint64_t size() const;
+    size_t size() const;
     
     void clear();
     
-    byte &operator[](uint64_t i);
+    byte &operator[](size_t i);
     
-    int growTo(uint64_t size);
+    int growTo(size_t size);
     
-    int growBy(uint64_t size);
+    int growBy(size_t size);
     
-    int quickShrink(uint64_t size);
+    int quickShrink(size_t size);
     
-    int quickRestore();
+    size_t quickRestore();
     
     bool isEmpty() const;
     
-    byte at(uint64_t) const;
+    byte at(size_t) const;
     
     int fill(byte v);
     
-    int fill(uint64_t start, uint64_t length, byte v);
+    int fill(size_t start, size_t length, byte v);
     
-    int fillFrom(const byte *input,uint64_t destStart,uint64_t len);
+    int fillFrom(const byte *input,size_t destStart,size_t len);
     
-    uint64_t append(const sp<_ByteArray> &);
+    size_t append(const sp<_ByteArray> &);
     
-    uint64_t append(const sp<_ByteArray> &, uint64_t len);
+    size_t append(const sp<_ByteArray> &, size_t len);
     
-    uint64_t append(const byte * data, uint64_t len);
+    size_t append(const byte * data, size_t len);
 
     sp<_String> toString() override;
     
@@ -72,12 +72,12 @@ public:
 
     static void Combine(ByteArray &dest,ByteArray appenddata);
 
-    static const uint64_t kDefaultSize;
+    static const size_t kDefaultSize;
 
 private:
     byte *mBuff;
-    uint64_t mSize;
-    uint64_t mPreviousSize;
+    size_t mSize;
+    size_t mPreviousSize;
     bool mMapped;
 };
