+++ util/concurrent/include/ThreadCachedPoolExecutor.hpp
@@ -33,9 +33,9 @@ DECLARE_CLASS(ThreadCachedPoolExecutor) IMPLEMENTS(Executor) {
 
     int awaitTermination(long timeout = 0) final;
 
-    int getPendingTaskNum() override;
+    size_t getPendingTaskNum() override;
     
-    int getExecutingThreadNum() override;
+    size_t getExecutingThreadNum() override;
 
     ~_ThreadCachedPoolExecutor() override;
