+++ util/concurrent/include/ThreadPoolExecutor.hpp
@@ -41,9 +41,9 @@ public:
 
     int awaitTermination(long timeout = 0) final;
 
-    int getPendingTaskNum() override;
+    size_t getPendingTaskNum() override;
     
-    int getExecutingThreadNum() override;
+    size_t getExecutingThreadNum() override;
 
     ~_ThreadPoolExecutor() override;
